refactor(android): extract snapshot hierarchy builder
